requirements.txt
demucs==4.0.1
inquirer==3.3.0
librosa==0.10.2.post1
python-reapy==0.10.0
Unidecode==1.3.8
yt-dlp==2024.8.6
ytmusicapi==1.8.0
# basic-pitch
